requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens @ file:///croot/asttokens_1743630435401/work
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm @ file:///croot/comm_1709322850197/work
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.57.0
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
docker==7.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.116.0
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1744912971218/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
itsdangerous==2.2.0
jedi @ file:///croot/jedi_1733987392413/work
Jinja2==3.1.4
joblib==1.5.1
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
kiwisolver==1.4.8
lightly==1.5.21
lightly-utils==0.0.2
lightning-utilities==0.14.3
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline @ file:///work/perseverance-python-buildout/croot/matplotlib-inline_1698864771271/work
mlflow==3.1.1
mlflow-skinny==3.1.1
mpmath==1.3.0
multidict==6.6.3
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.3.0
parso @ file:///croot/parso_1733963305961/work
pexpect @ file:///croot/pexpect_1750072593878/work
pillow==11.0.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
propcache==0.3.2
protobuf==6.31.1
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments @ file:///croot/pygments_1744664109463/work
pyparsing==3.2.3
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytorch-lightning==2.5.2
pytorch-ranger==0.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1734687138743/work
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentry-sdk==2.34.1
setuptools==78.1.1
six @ file:///croot/six_1744271502820/work
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.6.0+cu124
torch-optimizer==0.3.0
torchaudio==2.6.0+cu124
torchmetrics==1.7.3
torchvision==0.21.0+cu124
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.53.1
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.21.1
wcwidth @ file:///croot/wcwidth_1750352883074/work
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
zipp==3.23.0
